requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
mysql-connector-python==8.0.32
protobuf==3.20.3
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
